Created the file for executing a function every x steps
Say you have a column of 60 numbers, but you want to take the average of every 5 data points, so average 1-5, then average 6-10, and so on. This would do that! or you could sum them or whatever. You can use any function you want and any range you want.

The only stipulation is that they have to be in the same column though. Otherwise it wouldn't work like that. I may improve the functionality later on but this fit my needs for now at least.